+++ src/routes/completed/+page.svelte
@@ -3,4 +3,5 @@
 
 <div class="container">
 	<h1>完了しました</h1>
+	<div>お問い合わせありがとうございました。</div>
 </div>
