--- src/routes/form/completed/+page.svelte
@@ -1,4 +1,0 @@
-<script lang="ts">
-</script>
-
-<h1>完了！！！！</h1>
